Refactor BookList component into its own directory and add CSV export functionality. Moved styles to a new CSS module and created utility for boolean value conversion.
