+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import './App.css'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { Provider } from 'react-redux'
 import { store } from './redux/store.ts'
-import { BookList } from './components/BookList.tsx'
+import { BookList } from './components/BookList/BookList.tsx'
 import { YearSelector } from './components/YearSelector/YearSelector.tsx'
 
 const queryClient = new QueryClient()
