+++ frontend/src/components/BookList/BookList.module.css
@@ -1,0 +1,4 @@
+.buttonGroup {
+  display: flex;
+  gap: 8px;
+}
