+++ frontend/src/components/BookTable/constants.ts
@@ -1,0 +1,14 @@
+export const TABLE_HEADERS = [
+  'Title',
+  'Author',
+  'Rating',
+  'Pages',
+  'Date Start',
+  'Date End',
+  'Language',
+  'Format',
+  'Has Smut Scenes',
+  'Book Club Choice',
+  'Queer',
+  'First Time Read'
+]
