+++ frontend/src/utils/getBoolValue.ts
@@ -1,0 +1,8 @@
+/**
+ * Returns 'Yes' if the value is true, otherwise returns 'No'
+ * @param value - The boolean value to check
+ * @returns 'Yes' if the value is true, otherwise 'No'
+ */
+export const getBoolValue = (value: boolean | undefined) => {
+  return value ? 'Yes' : 'No'
+}
